Modified Code for Loading HipMRI data and Test
Code provided in the task sheet was not working, so the code used in utils.py is slightly different - it checks for each image's dimensions and updates the max number of rows*columns possible for an image.

It appears that the training and testing datasets contain images of different sizes. The output of loading both datasets is also provided in s4878126_VQVAE.out.