Skeleton for Embedding Layer Added.
Note: previous commit was actually an overwritten copy. Not realising that I was not editing on rangpur, but instead a local copy, I copied the rangpur version and overwrote the latest changes to the code. Need to pay more attention in the future as to which file I am actually editing in vscode.