Implimented basic model training code + fixes
Implimented basic model training process eg
running through epopchs and training the model and doing a
quick accuracy test at the end. Still needs hyperparimeter tuning
as well as likely some changes inside the model e.g. more/wider
layers.

Also had to fix some bugs with the output of the model as it was
not in the correct format needed for the loss function