Changes to VisonTransformer/Transformer bug fixes
VisionTransformer was not behaving properly
due to the setup of the multi head attention
layer in the TransformerEncBlk, this was fixed by
togelling the batch_first option to True as the inputed
data into this layer is batched. Therfore this option
needed to match that.

Additionally some small changes were made to the vision
transformer with the setup. The main change was introducing
a variable amount of transformerEncBlk items. All other
small changes were to support this change / clean up code.