Finished CvT implimentation - MHA might not work
Have finished CvT Implimentation by making the equivalent
modules used for the original ViT but converting them to allign
with the CvT Architecture.

I'm not sure that the pytorch MHA is doing the correct thing,
as the information for the CvT's mha seems to have a few
nuance steps, so that might be causing some issues.

Performance on the first through run throughs isn't great so nothing
to report on that front