Added global filter layer and increased model complexity and epochs for better performance
Previous training with 50 epochs and a learning rate of 0.00001 resulted in much lower training accuracy than expected. The strategy to deal with that was to add a global filter layer of 1x1 for the refinement of features and increase the number of epochs for deeper training. This extra global filter will allow it to capture global dependencies from the feature maps. It is expected that this will let the model learn features of higher complexity that would result in better overall accuracy.