Completed training with 200 epochs, achieving over 91% accuracy on both training and validation set
It can be observed from the following that significant improvements have been achieved by the GFNet model in training and validation after training 200 epochs with a learning rate of 0.001. The final value of training accuracy rises to 91.91%, while the validation accuracy of the model has risen to 91.46%. While the training loss has reduced to 0.1209, the reduction of validation loss to 0.0583 shows that the generalization of the model is pretty good with minor overfitting.